Do not export SingularityCmd (#4541)
This is a minimal change to avoid exporting SingularityCmd so that we
don't have to chase its uses all over the place. In order to avoid
touching the few instances outside
github.com/sylabs/singularity/cmd/internal/cli where SingularityCmd is
being used, introduce a couple of functions in that package to provide
access to this command.

There's one used to generate bash completions, where it doesn't really
need access to the command, it just needs a way to request the cobra
bash completions to be generated.

There are a couple of other ones where they need to traverse the command
tree to access every command, and SingularityCmd is just the starting
poont. While the change is still leaving the door open for arbitrary
modifications, it makes it easier to track down these accesses.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>