Do some minor CLI cleanup
Even if a lot of lines are changing, there are not significant changes
here. It's mostly about removing duplicated flag definitions (multiple
flags are touching the same package-level variable) and moving them to a
more neutral location (e.g. force is defined in build.go but it's used
in all the actions).

Split runBuild into runBuildLocal and runBuildRemote.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>