Bubble context.Context as far up as possible
There are several cases of a function requiring a context.Context, and
the code creating said context just before calling that function.

For context to be effective, it should be created as close as possible
to the request, and singularity being a CLI application, this means as
close as possible to the CLI UI as possible.

This change was created by locating all the context.TODO() and
context.Background() calls, removing them and converting the context to
a variable passed to the corresponding function, until hitting either a
function that was already receiving a context.Context or the CLI. Sadly
this had quite a bit of splash damage, as it touched a couple of
functions in the engine, which required changing the interface
definition, which required touching the other engine implementations,
which required touching other code.

There are no real functionality changes here, only whatever is needed to
introduce the context parameter wherever it's needed.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>