Split TestSquashfs
One test is packing too much things in it. There are three separate
tests here, so there should be three separate tests.

Split TestSquashfs into three separate tests.

Add another one to cover the case where mksquashfs runs but exits with
an error. Given the weird API, there should be a test for the case where
the user sets MksquashfsPath to something other than a valid mksquashfs
program, but that's not really testing the code anymore.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>