Error if '--sif-id' is 0
If the user passed `--sif-id 0`, sinse that is the default
`verify`/`sign` behavior, it will seem like that flag has no effect
(witch it does not), it should fail as an invalid value.

There are probably better ways to solve this, but this is what I came up
with. Open to suggestions.