Update goversion to require Go 1.13
The package github.com/sylabs/singularity/internal/pkg/util/goversion
provides a build-time check for the Go version being used. This is
mirrored in mconfig, where the required version of the compiler is also
specified.

2fe5a412810d01f1c03774cec9c47ba5c0f8e453 updated most of the locations
where a Go version is specified, but missed this one, sorry.

This change makes it a build-time failure to try to build singularity
with something older than Go 1.13.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>