Add run-linter script
This wraps all the ugly logic around (optionally) downloading and
verifying that the golangci-lint version is the correct one.

This does introduce strict checking for the golangci-lint version
(1.20.0 in this case), as it has been observed that the results change
from version to version. But it also leaves the window open for people
to use a locally built golangci-lint, which reports its version as
"(devel)".

People wanting to control the location of golangci-lint can download and
install it themselves, prior to running this script.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>