Fall back to current working directory outside git working copy
When running run-linter in a release directory (as opposed to a git
working copy) the _intention_ is to fall back to the current working
directory, with the assumption that the user is calling something like
"scripts/run-linter".

During refactoring this was removed and the code is no longer working
with a release tarball. Fix this.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>