Remove vendor directory and associated flags
Align with the rest of the singularity repositories and DO NOT use a
vendor directory. Instead rely on the Go proxy to preserve versions of
the modules we need.

Rename scripts/check-vendor to scripts/check-go.mod. Instead of checking
the vendor/ directory, check the state of the go.mod and go.sum files.

Set GOPROXY to https://proxy.golang.org (the default value is
"https://proxy.golang.org,direct") to work around issues with _ancient_
versions of Git shipped with some distributions. It should be possible
to override this using normal make constructs.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>