Make sure we use the same go binary in go-generate
This is hopefully the last case of a naked "go" invocation. The rest of
the go calls are happening using the path that mconfig determined, and
this change is fixing that for the _program_ that go-generate ends up
running.

When you run "go generate" (what "go-generate" does) it looks for
'//go:generate' comments and runs whatever command is specified there in
order to generate something. This goes thru the shell, so it's OK to use
environment variables on that command line.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>