Upgrade github.com/containers/image to a released version
For some reason dependabot is not picking this one up (is it respecting
the fact that the module is pinned to a commit?). Upgrade
github.com/containers/image to a released version, instead of the commit
we have pinned that module to. That brings github.com/containers/storage
with it.

Fixes: #4648

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>