Get plugin manifest in a separate process
We need to get plugin's manifest in a separate to process to avoid an
error when we are compiling a plugin which is already installed. If a
plugin is already installed and enabled, singularity will have already
loaded it by the time we get to retrieve the manifest from the new
version, and that fails because retrieving the manifest requires loading
the plugin again.

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>