Ask for the version number only once
Currently the script "make-dist" calls "get-version", which is also
called in "mconfig". The problem is that mconfig has additional logic
overriding the output of get-version (to support use cases where the
resulting version number is not singularity's version number).

When generating the .spec file, versions like "3.5.0-rc.1" get
transformed into "3.5.0~rc.1" so that rpm considers "3.5.0-rc.1" an
older version than "3.5.0", matching the semver interpretation of that
version number. Versions like 3.5.0+1-abcde are transformed to
3.5.0+1.abcde because rpm doesn't like "-" in the version number.

To avoid making too many changes in (what should be) disconnected
places, have mconfig call get-version, and if the user wants to replace
that with something different by using the -V flag, simply override the
resulting version number and pass that to make-dist as an argument.
make-dist is then called with the correct value when the user runs "make
dist", and writes it to the VERSION file in the release tarball.
get-version will _read_ that file and output whatever the user passed as
argument to -V. This causes the "singularity version" command to report
the actual version number (as recorded in the VERSION file) instead of
some fake x.y.z one, and this helps us to support users.

In this way everything stays consistent.

Follow-up-for: 15bbd6153189a9949a2814c625794d8839021a0d
Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>