Restore vendor/ for release tarballs
In order to be more friendly to people wanting to build singularity in
air-gapped or air-gapped-like situations (like Linux distributions),
restore the vendor directory in release tarballs (and only in that
case).

There's a simple assumption here: if there's a vendor directory, we must
pass -mod=vendor, and if there isn't, we pass -mod=readonly to make sure
that we are not unintentionally adding modules.

Since this affects the resulting binary interface, plugins need to be
aware of this. Make the corresponding adjustments there.

The vendor directory is created when running "make dist". To avoid
making larger changes, make it a requirement to use the make-dist script
from the makefile. It's still possible to use it directly if the user
runs "go mod vendor" manually _before_ calling scripts/make-dist.sh. In
the future all the scripts requiring access to the Go tool path, build
flags, etc, can move to sourcing a shell file with those settings.

Fixes: #4664
Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>