Information / option to fix restrictive sandbox permissions
Fixes: #4628

With umoci unpacking, 3.5 will honor the permissions in OCI/docker
layers when extracting them. This can result in a sandbox build
containing files/dirs that cannot be 'rm'd without modifying the
permissions. To make things easier for users, and provide a way to use
the 3.4 behavior in a transition period:

- Add a `--fix-perms` option that will modify the permissions on
oci/docker layer extraction like 3.4 and earlier did.
- Warn the user about future removal of the `--fix-perms` option, and
provide a link to the discussion issue.
- When `--fix-perms` is not used, scan the result of OCI/docker extraction
to warn the user if any files in there will not be easily removable,
directing them to `chmod -R u+rwX` and the `--fixperms` option.