Fix rocmliblist for OpenCL apps and blender example
In order for a blender OpenCL example container to work we need to bind
in things from /opt/rocm/opencl/lib via `--rocm`

We also need to uncomment the libpci, libelf, libnuma binds as though
all the ROCm GPU computer containers I've tried work, the blender OpenCL
example trips up with these not available in the container.

Now blender works for OpenCL rendering with `--rocm --bind /etc/OpenCL`
(the bind is needed for the vendor icd) which is not needed for general
ROCm workloads.