Fix "build_and_test" workflow failure on push
In CircleCI we (unintentionally) configured the build_and_test workflow
to run both on pull requests and pushes, and in the later case the
workflow is failing because no pull request URL is available.

Fix this by handling the case explicitly and assume that pushes MUST run
E2E tests. In that way, even commits that didn't require E2E to run will
get the benefit of testing (after being merged).

Signed-off-by: Marcelo E. Magallon <marcelo@sylabs.io>