Fix disk space leak when building from docker-archive
Fixes #4560

When building a singularity .sif from a docker archive, singularity would
create a temporary file named docker-tarXXX in /var/tmp, and never clean it up.
This adds a line to calculateRefHash (which was creating the file via
NewImageSource()) to close the image source it created, cleaning up the
temporary file.