Use nil pointer for noop DockerAuthConfig
Fixes #4860

In #4544 some tidy up resulted in our DockerAuthConfig being
dealt with by value, not as a pointer. Unfortunately in
`containers/image` the test of whether to use the DockerAuthConfig
is against `nil`, and not a default/empty DockerAuthConfig.

We need to pass a nil pointer into `containers/image` routines, or it
will not fall back to other authentication sources, such as the
`.docker/config.json` file.