Correct nc args to fix TestAddDelNetworks with OpenBSD nc
Fixes #4894

In TestAddDelNetworks we are using `nc` to listen on port 80.

We are using the command `nc -l -p 80`. On a strict openbsd version of
`nc` this is not permitted. `-p` is the `source-port` and not valid for
using nc to listen. `-l` with `-p` happens to work on traditional and
patched versions of `nc`, but the correct options that work across all
versions are `nc -l 80`.