Kill helper process in network_linux_test on error
If the network setup fails with an error, our helper process such as nc,
may be stuck. E.g. in the tests using nc we are expecting it to end once
a connection is made and closed. If the network setup failed we need to
kill it explicitly to avoid a freeze.