Handle non-standard awk location
Under nix, guix, and maybe some other environments there is no
`/usr/bin/awk` - it has to be found on `PATH`.

The GNU `-S` option to `/usr/bin/env` allows use of shebangs with multiple
arguments passed to the interpreter (as is the case for genmod.awk),
but this is non POSIX so won't work under busybox etc.

To support bare POSIX environments as well as nix type environments the
`genmod.awk` can be made into a polygot program where some valid `sh`
calls the script again using `awk` on `PATH`.

Ref: https://unix.stackexchange.com/a/149801