Nicer encryption error messages
Fixes: #5058

Stop using github.com/pkg/errors Wrap in the encryption code, as this
adds a stack trace to error output for common issues such as an
incorrect key being supplied. Also try to make some of the error message
a bit simpler.

We can use Go 1.13 %w when we want to have nice wrapped errors, but for
now avoid this and use %v as RHEL8 doesn't have 1.13 as an app stream
yet.