Use cp -H symlink behaviour in multi-stage Copy
Fixes: #5059

In a multi-stage build up until 3.5.3 we would follow all symlinks when
copying files, using cp -L. This was changed so that symlinks were not
followed in 3.5.3 via #4873.

In #5059 it was noted this breaks some NVIDIA container make builds.
Also, Docker will follow symlinks in multi-stage build COPY if they
are specified directly, or via glob. It will not follow symlinks that
are copied as a result of directory traversal. This is the cp -H
behaviour, and it makes sense to follow it for least surprise vs Docker
multi-stage builds.