Use base metapackage for arch bootstrap
Fixes: #5077

We used to pull the 'base' group for arch, and then filter out some
packages listed in it, installing the remainder via a pacstrap
bootstrap. This no longer works as the 'base' group has been removed
and there is now a 'base' metapackage:

https://www.archlinux.org/news/base-group-replaced-by-mandatory-base-package-manual-intervention-required/

Because this is a single package install for pacstrap there isn't a
nice way to filter out the things that we weren't installing into the
container with the previous method. This gives larger container
images, but comparable to the dockerhub official arch images.

The alternative would be to install things listed as dependencies of
'base' by keeping track of that manually here, or writing complex
query / exclude code. I feel these are probably a bit too brittle, and
risk breakage later on while simply using the `base` metapackage
should be okay. Have left the package list as a string slice in case
we want to go back to a list of packages to install in future.