Add --no-hostfs actions flag
Closes: #5081

Add a --no-hostfs so that if the global config specifies "mount hostfs
= yes" it is still possible to run a container without all the host
filesystems mounted.

An example of how this is useful - for containers that have
directories with generic names like `/data` inside the
container. `/data` is a mount point on my host so it's useful to be
able to turn of `mount hostfs` for these containers.