Bump umoci to commit with xattr error suppression
Fixes: #4985

PR https://github.com/openSUSE/umoci/pull/311 was merged upstream. This
is our patch to stop umoci warning for every individual xattr error when
extracting to an FS without xattr support. Was hoping to point to a
release containing this, but there isn't a release yet and we do want
this patch in 3.6 as without it output is very noisy.