Process appinstall sections in order in build
Fixes: #4820

SCIF apps can be written so that in a definition file the %appinstall
section for the second SCIF app in the def file depends on things
created in the %appinstall for the first SCIF app. Singularity is
storing def file sections in un-ordered go maps, so %appinstall sections
weren't run in the order they were parsed from the def file.

This PR tracks the order of SCIF apps as they are read from the def file
so that they dependent %appinstall sections will work as expected.