Raise `max loop devices` in e2e/config test
 - Fixes: #5370

When we are on a system with a lot of cores, and/or loop devices are
consumed by snaps, flatpak apps etc. then the test setting of `max loop
devices = 10` may not be sufficient to find a free loop device.

Setting it to `128` which is 1/2 of the default. Resolves issue running
tests on a 48 thread machine with Fedora 32 / VM Ubuntu 20.04.