Skip inaccessible loop devices in shared loop code
 - Fixes: #5380

It's possible for a loop device to become inaccessible under some
circumstances that have been reported to us, such as swapping out a
symlinked image that resides on AFS.

Instead of this causing a fatal error when using shared loop devices,
skip over the inaccessible loop device and move on to consider the next
one.

Also adding a debug statement when we find a shared loop device to use
to make that more obvious in any bug reports.