Update golangci-lint and make version check more permissive
Closes #5597

The restriction to using only the old 1.21.0 version of golangci-lint
for make check is inconvenient, as developers are unlikely to have that
version on their system.

Update to the latest version, and make the check for golangci-lint
verion permissive. Warn instead of error if the version doesn't match
the one we will install, and is used in CI.