Restore umask in container, actions, not fakeroot
Closes #5214

When we are running through an actions flow, and are *not* using
fakeroot then by default save the umask from the calling environment and
set it within the container.

This ensures that when a user works inside a container then files they
create on host filesystems have the permissions that they expect, given
their umask.

Add a flag `--no-umask` to disable this propagation.