Don't install old golangci-lint in individual CI jobs
Fixes #5634

In recent PR #5609 scripts/run-linter was updated to use the latest
golangci-lint.

However, it was missed that some of the CI jobs are installing their
own 1.21.0 version explictly, so some checks are run with 1.21.0 and
some with 1.31.0.

Remove the install of 1.21.0 so everything pulls the right linter via
`scripts/run-linter` in the make check target.