Give useful message when we can't extract an ext3 image
 - Fixes #5385

In unpriv mode we cannot loopback mount an old style ext3 container
image in order to extract it into a temporary sandbox.

Currently we are exiting with an error that the image is "not a squashfs
root filesystem". This is not very friendly to users / doesn't hint at
the resolution.