Force a prompt if `PS1` is empty entering a container.
 - Fixes: #5340

If a container without a `.singularity.d` is run then there is no
environment file to set a custom `$PS1`.

If additionally singularity is called through a script, which will reset
`$PS1` to empty, then a confusing silent prompt will result.

In our action script force PS1 to a default if it is empty.