Allow /etc/localtime & /etc/hosts bind to fail
 - Fixes: 5465

In `singularity.conf` and hard-coded under `--contain` we bind
`/etc/hosts` and `/etc/localtime` to ensure sensible timezone and
hostname resolution in the container.

In certain circumstances the container may be such that the bind mount
onto `/etc/hosts` or `/etc/localtime` fails. This is a fatal error, but
it would be nicer just to warn, as the container is likely still usable.