Ensure files staged into build do not match distro package content
 - Fixes #5250

During builds we stage copies of `/etc/hosts` and other files on the
host system, and bind these copies into the container for `%post`.

On RHEL/CentOS/Fedora, when `/etc/hosts` on the host system is
unmodified from the default content provided by the distro's 'setup'
package, yum will try to rename & replace it if the 'setup' package
is reinstalled / upgraded. This will fail as it is bind mounted, and
cannot be renamed.

To work around this we just add a newline to the end of files we are
copying and binding in build.createStageFile. Adding a newline means
the staged file is now different than the one in the 'setup' package
and yum will leave the file alone, as it considers it modified by the
user.