Try /sbin/ldconfig for gpu libs if $PATH ldconfig fails
Fixes: #5002

On systems with Nix/Guix type userland on top of a 'standard'
distribution there may be a non-working userland `ldconfig` on `$PATH`
that will fail to find the system installed CUDA libs. If we fail to
execute `ldconfig` on `$PATH` then fall back to trying the default
`/sbin/ldconfig` in case that works.