unsquashfs - don't extract xattrs where not supported by fs
Fixes: #5668

Corrects an issue that mainly occurs on Fedora 33 / Ubuntu 20.04 to a
tmpfs, as they use unsquashfs 4.4 which returns error codes for
non-fatal errors.

First we try unsquashfs with appropriate xattr options.

If we are in rootless mode we need "-user-xattrs" so we don't try to
set system xattrs that require root.

However...

 1. This isn't supported on unsquashfs 4.0 in RHEL6 so we have to fall
    back to not using that option on failure.

 2. Must check (user) xattrs are supported on the FS as unsquashfs >=4.4
    will give a non-zero error code if it cannot set them, e.g. on tmpfs
    (#5668).