Overhaul `singularity search`.
 Fixes: #4438

Search for the current architecture, or alternate with --arch flag.

Limit search to signed containers with the --signed flag.

Don't display matching users or collections - these aren't really
useful.

For matching images show tags, or if no tag the hash that can be used to
pull.

Show description for an image if set, and fingerprints if it is signed.