Fix progress bar display when content-length unknown
 - Fixes #5826

This fixes an issue where if an HTTP server doesn't return
a content-length the download progress bar shows meaningless `-1.0b`
`0%` `-5s` size / percentage / ETA labels.

We rewrote progress bar handling in 3.7 to use a different package, and
this was missed.

I had to update to mbp/v6 here (from v4) as v4 doesn't have the ability
to show the amount downloaded without the total.