e2e: Force insecure policy for skopeo
Skopeo is used to copy docker://busybox into the local e2e oci registry
for e2e tests. This is failing as an updated version of skopeo requires
an /etc/containers/policy.json but none is provided by the skope apk.
Can avoid the requirement with `--insecure-policy` per @cclerget.

Fixes #5882