Disable -race for tests on ppc64le only
Closes #5914

On ppc64le the e2e tests will fail if run with -race under Go 1.14 or
1.15.

They worked correctly with Go 1.13, but newer versions cause the CGO
init() method to be called twice, the latter time unprivileged... so
it fails. Not clear why this happens. Only on ppc64le, not amd64 or
arm64.