Remove TestSelfTest integration test
`singularity selftest` was a command in previous versions of
Singularity. A test for it was ported over 3 years ago, but the
command never was. The test that is being removed here expects
`singularity selftest` to fail. It doesn't need to go across to the
e2e suite, as I don't believe there are any plans to implement
`singularity selftest` in 3.x.