Do not singularity build SIF in image unit tests
The pkg/image unit tests were using `singularity build` to create a
SIF on which to test the image code. We do not want unit tests to call
out to Singularity - they should ideally run without Singularity
installed.

Use the image we have on hand for our e2e tests.