Remove action-test target
The `action-test` target was running the obsolete
TestSingularityActions integration test code as part of the rpmbuild
CI steps. That test code is now removed, as it is obsoleted by the
e2e/actions and is unmaintained.

Added a simple `singularity exec` 'sanity check' to the rpmbuild CI
step in its place.