Remove cmd/singularity/build_test.go & deffile_test.go
Have gone through the code and verified that:

1. imageVerify is covered by reimplementation in
`e2e/internal/e2e/image_verify.go`.

2. `TestBuild` cases are covered in `e2e/imgbuild`. The cases are
covered in 'buildFrom' and image verification is equivalent.

3. `TestBadPath` is covered in `e2e/imgbuild` via `badPath`

4. `TestMultiStageDefinition` cases are covered in `e2e/imgbuild` via
`buildMultiStageDefinition`

5. `TestBuildDefinition` is covered in `e2e/imgbuild` via
`buildDefinition`. The custom image verification has been ported to
`DefinitionImageVerify` in `e2e/internal/e2e/image_verify.go` and is
equivalent.

The remaining area that is not directly covered 1:1 in the `e2e` suite
is `TestMultipleBuilds`. This tests chained builds, where the output
of one becomes the input of another. The paths tested are...

* (Library URI -> SIF) -> SIF
* (Library URI -> Sandbox) -> SIF
* ((Library URI -> SIF) -> SIF ) -> SIF (with labels via def file)
* (Library URI -> Sandbox) -> SIF (with labels via def file)

In the e2e tests we have `buildLocalImage` which tests:

* SIF -> SIF
* Sandbox -> SIF
* SIF -> SIF (via def file)
* SIF -> SIF (with labels via def file)
* Sandbox -> SIF (with labels via def file)

In conjunction with the other test in the `e2e/imgbuild` code which
cover the `library://` to SIF steps, and other def file combinations I
am happy that we are covering all of the build steps of
`TestMultipleBuilds`.