Streamline makefile test rules
Get into a simple `make test` and `make testall` situation such that:

- Make `test` run all short tests that don't need docker pull
etc. (`short-unit-test` && `short-integration-test`)

- Make `testall` run the current full suite of tests equivalent to a
CI run. Will require a docker auth env var setup to complete
properly. Uses the `short-unit-test` target instead of `unit-test`
until unit test deduplication is complete and/or docker auth is
respected in those. Retain the `check` target so it also enforces the
linter as CI does.

Ensure that the integration tests have short and normal targets that
match the remaining network tests. For now, they do identical things,
but let's stick to the short / normal pattern as there is likely
'unit' test code that really should be considered an integration test,
and may be refactored in future.