oras: Work around harbor failure to accept SifConfigMediaType
Closes #5691

ORAS CLI and code examples generally push an ORAS artifact as a layer of
specific layer media type, and with a null config of a specific config
media type. The majority of registries accept this.

Harbor does not accept the null config that has our specific
SifConfigMediaType. However, it will accept an oras Push() without the
config set at all.

We do want to push up the null SIF config media type as this is the
canonical method in the oras docs, and e.g. results in a nicer error
if you mistakenly try to pull the image with an OCI container engine like
docker.

Since the error we get back from harbor not accepting the config is so
generic, perform a simple retry without a config when there is an
initial failure.

Also separately fix our config media type per the oras docs by removing
the `+json` extension as we only ever send a null config.