Implement a copy-through ProgressCallback for silent loglevels
Fixes #5924

A nil ProgressCallback is returned for silent / quiet loglevels which
leads to a panic when this is blindly used by image download clients
(shub / net).

To avoid having to check for a nil callback and behave differently in
the client code, return a callback that does a straight copy-through in
these cases.