fix: wait for cryptsetup /dev/mapper device to appear (#5953)
The device /dev/mapper/... does not show up before cryptsetup returns. This creates a race condition as, if singularity attempts to mount /dev/mapper/..., before it shows up, it fails.

This pull request causes singularity to wait up to 25.5s for /dev/mapper/... to show up, through a series of exponential back off sleeps, before continuing. If it doesn't show up in that time it will abort with an error.

Fixes #5950