Do not show keystore URI in newpair confirmation message
The real URI to which we are pushing is computed at a lower level, in
remote endpoint handling, than we have access to in the CLI newpair
code. At present it is only correct for the default keystore.

Remove the URI from the message, until we can address the issue by
having the CLI code able to know the true correct URI in all cases.

Fixes #4

See also #23