Some updates to make the debian packageing work.
Because the directory was moved to the subdir dist, it is needed
to copy the directory to the top directory or set a link
before running debuild.

As usual the command for building the debian package is:
debuild -b -us -uc

To make it configurable some environment variables can be set.
Due to the fact, that debuild filters out some variables, all the
configuration variables need to be prefixed by DEB_

options for mconfig, by setting the environment variables:
DEB_NOSUID=1    adds --without-suid
DEB_NONETWORK=1 adds --without-network
DEB_NOSECCOMP=1 adds --without-seccomp
DEB_NOALL=1     adds all of the above

To select a specific profile for mconfig:
DEB_SC_PROFILE=release-stripped
DEB_SC_PROFILE=debug

The builddir can be specified with this:
DEB_SC_BUILDDIR=builddir

Currently no other config options can be set via environment,
but it would be easy to add them if needed.

I used as much as possible from the existing rules file for this
first step.

In the control file I added a build dependency to golang and changed
the architecture to amd64.