Inherit Docker/OCI LABELs during build
When building / pulling from a Docker/OCI source, retrieve LABELs from
the config, and ensure they are set in the resulting Singularity
container.

This fixes a longstanding regression from Singularity 2.x. Note that
you will now need to use `--force` in a build to override a label that
already exists in the source Docker/OCI container.

Fixes sylabs/singularity#146