Handle absolute symlinks in multi-stage copy
In a multi-stage build `%copy`, if the source for the copy line is a
symlink we should dereference it and copy the target. Note that if we
are copying a directory we shouldn't dereference any symlinks
within the directory, only the directory itself.

Previously we were doing a straight `cp -frH` to implement this, but
it does not handle the case where a symlink to be copied is an
absolute symlink. The absolute symlink's target is correct in the
context of the stage's rootfs, but we are copying in the host context.

To fix this we manually dereference the top level src of a `%copy` line,
in a manner that will resolve absolute symlinks relative to the src
stage's rootfs correctly.

This PR also:

 - Splits host->rootfs copy, and stage->stage copy functions.
 - Rewrites the tests for host->rootfs and stage->stage copy
   functions to cover more cases including top-level and nested source
   files, dirs, and absolute + relative symlinks.
 - Uses securejoin to restrict resolution of paths to being inside of a
   rootfs where this is appropriate. This is for least-surprise rather
   than security, as the hostfs is still modifiable in `%setup` etc.

Fixes #159

Conflicts:
	go.mod