Add `--compat` short-hand for improved OCI/Docker compatibility
It is common for users to run docker containers that expect more
isolation than is default in Singularity, and that can create files on
startup. This is a simple short-hand to enable `--contain-all,
--no-init, --no-umask, --writable-tmpfs`. These options give the best
chance of an OCI/Docker container working as expected but *without*
requiring the user/uts/net namespaces that we can't rely on in all
installations / configurations of Singularity.

Fixes: sylabs/singularity#75