Change log-plugin example to run at CLI level
The example log-plugin was logging a message to syslog when a
container was run, from a PostStartprocess callback. This has the
disadvantages that:

- It only logs container executions, not other operations that might
  warrant auditing, e.g. pushing containers to external locations.
- The friendly URI/image path of a `singularity run docker://alpine` or
  similar is not available, as only the true SIF file path is present in
  the engine configuration that can be accessed by the plugin. This may
  be an obfuscated path in the cache.

hpcng/singularity chose to add the URI/SIF argument into the engine
config so a PostStartprocess callback can access it.

Instead, rewrite the plugin so that it is an example that runs as a
PreRunE on the root `singularity` command, can log all actions, and
has access to log the command args that include the friendly URI of a
container etc.

Fixes #119